pyproject.toml
[project]
name = "kroos-indexation"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "psycopg2-binary>=2.9.9",
    "qdrant-client>=1.9.0",
    "openai>=1.30.0",
    "pdfplumber>=0.11.0",
    "openpyxl>=3.1.0",
    "pandas>=2.2.0",
    "python-docx>=1.1.0",
    "python-dotenv>=1.0.0",
    "schedule>=1.2.0",
    "tiktoken>=0.7.0",
    "anthropic>=0.25.0",
]
